requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.3
pydantic==2.9.2
pydantic-settings==2.6.1
marshmallow==3.22.0
marshmallow-sqlalchemy==1.0.0
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.1
validators==0.22.0
apscheduler==3.10.4
sqlalchemy==2.0.34
mysqlclient==2.2.4
mysql-connector-python==9.1.0
langchain==0.3.9
langchain-core==0.3.29
langchain-community==0.3.8
langchain-text-splitters==0.3.3
langchain-chroma==0.1.4
langchain-experimental==0.3.3
langchain-openai==0.3.0
langchain-google-genai==2.0.0
chromadb==0.5.3
pypdf==5.0.0
beautifulsoup4==4.12.3
docx2txt